requirements.txt
Django==4.2
easyocr
torch
google-generativeai
